requirements.txt
numpy
scikit-image
scikit-learn
joblib
python-multipart
pillow
fastapi
matplotlib